requirements.txt
tornado==4.5.2
pymongo==3.5.1
firebase_admin==2.9.1
requests==2.18.4
bcrypt==3.1.7
raven==6.6.0
simplejson==3.13.2
google-cloud-core==1.0.2
google-cloud-firestore==1.3.0
protobuf==3.9.0
grpcio>=1.8.2
beautifulsoup4==4.7.1
python-slugify
pyjwt
tornado-redis
mailchimp3
stripebcrypt==3.1.7
beautifulsoup4==4.7.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
firebase-admin==2.9.1
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.0.2
google-cloud-firestore==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.6
mailchimp3==3.0.14
protobuf==3.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.5.1
python-slugify==4.0.1
pytz==2020.1
raven==6.6.0
requests==2.24.0
rsa==4.6
simplejson==3.13.2
six==1.15.0
soupsieve==2.0.1
stripe==2.50.0
text-unidecode==1.3
tornado==4.5.2
tornado-redis==2.4.18
urllib3==1.22
